Allow usage of visualCaseGen on unported machines
Users should now be able to run visualCaseGen on machines where
CESM is not ported. The entire configuration workflow should work
except for the case creation stage (final step). Instead,
users can print out all the necessary steps to replicate the configuration
on a supported machine.

+++ tests/3_system/test_custom_mom6_grid.py
@@ -156,17 +156,27 @@ def test_custom_mom6_grid():
 
         cvars["PROJECT"].value = "12345"
 
-        # *Click* the create_case button
-        safe_create_case(cime.srcroot, case_creator)
+        try:
+            # *Click* the create_case button
+            safe_create_case(cime.srcroot, case_creator)
+
+            # sleep for a bit to allow the case to be created
+            time.sleep(15)
         
-        # sleep for a bit to allow the case to be created
-        time.sleep(15)
-        
+        except RuntimeError as e:
+            if "not ported" in str(e):
+                print("CESM is not ported to the current machine. Skipping case creation.")
+            else:
+                # If the error is not related to machine porting, raise it
+                raise e
+
         # remove mom6_bathy notebook belonging to the test_grid:
-        os.remove(nb_path)
+        if os.path.exists(nb_path):
+            os.remove(nb_path)
 
         # remove the caseroot directory
-        shutil.rmtree(temp_case_path)
+        if os.path.exists(temp_case_path):
+            shutil.rmtree(temp_case_path)
 
 
 if __name__ == "__main__":
