+++ tests/3_system/test_f2000_custom_grid.py
@@ -173,14 +173,23 @@ def construct_custom_res_from_std_grids(cime):
 
         cvars["PROJECT"].value = "12345"
 
-        # *Click* the create_case button
-        safe_create_case(cime.srcroot, case_creator)
+        try:
+            # *Click* the create_case button
+            safe_create_case(cime.srcroot, case_creator)
         
-        # sleep for a bit to allow the case to be created
-        time.sleep(15)
+            # sleep for a bit to allow the case to be created
+            time.sleep(15)
         
+        except RuntimeError as e:
+            if "not ported" in str(e):
+                print("CESM is not ported to the current machine. Skipping case creation.")
+            else:
+                # If the error is not related to machine porting, raise it
+                raise e
+
         # remove the caseroot directory
-        shutil.rmtree(temp_case_path)
+        if os.path.exists(temp_case_path):
+            shutil.rmtree(temp_case_path)
 
 
 def construct_custom_res_from_modified_clm_grid(cime):
